+++ tukaan/dialogs/__init__.py
@@ -1,7 +1,8 @@
-from .about import AboutDialog
-from .calendar import Calendar
-from .colorchooser import ColorChooser
+from .aboutdialog import AboutDialog
+from .datetimedialogs import DatePicker, TimePicker
+from .colorpicker import ColorPicker
+from .fontpicker import FontPicker
 from .filedialog import FileDialog
-from .input import InputDialog
+from .inputdialog import InputDialog
 from .messagebox import MessageBox
-from .notifications import Notification
+from .notification import Notification
