+++ tukaan/dialogs/colorpicker.py
@@ -1,0 +1,7 @@
+from __future__ import annotations
+
+from .dialog import Dialog
+
+
+class ColorPicker(Dialog):
+    ...
