+++ tukaan/dialogs/datetimedialogs.py
@@ -3,5 +3,9 @@ from __future__ import annotations
 from .dialog import Dialog
 
 
-class ColorChooser(Dialog):
+class DatePicker(Dialog):
+    ...
+
+
+class TimePicker(Dialog):
     ...
