+++ tukaan/dialogs/fontpicker.py
@@ -3,5 +3,5 @@ from __future__ import annotations
 from .dialog import Dialog
 
 
-class Calendar(Dialog):
+class FontPicker(Dialog):
     ...
